+++ moq-relay-ietf/src/main.rs
@@ -20,6 +20,7 @@ pub use web::*;
 
 use std::net;
 use url::Url;
+use moq_transport::session::SharedState;
 
 #[derive(Parser, Clone)]
 pub struct Cli {
@@ -69,6 +70,9 @@ async fn main() -> anyhow::Result<()> {
         anyhow::bail!("missing TLS certificates");
     }
 
+    let shared_state = SharedState::new();
+    let relay_stopping_state = SharedState::new();
+
     // Create a QUIC server for media.
     let relay = Relay::new(RelayConfig {
         tls: tls.clone(),
@@ -76,7 +80,9 @@ async fn main() -> anyhow::Result<()> {
         node: cli.node,
         api: cli.api,
         announce: cli.announce,
-    })?;
+    }, shared_state.clone(),
+       relay_stopping_state.clone(),
+    )?;
 
     if cli.dev {
         // Create a web server too.
@@ -84,8 +90,10 @@ async fn main() -> anyhow::Result<()> {
         let web = Web::new(WebConfig {
             bind: cli.bind,
             tls,
-        });
-
+            shared_state: shared_state.clone(),
+            relay_stopping_state: relay_stopping_state.clone(),
+        }
+    );
         tokio::spawn(async move {
             web.run().await.expect("failed to run web server");
         });