+++ moq-relay-ietf/src/session.rs
@@ -1,6 +1,7 @@
+// session.rs
 use futures::{stream::FuturesUnordered, FutureExt, StreamExt};
 use moq_transport::session::SessionError;
-
+use moq_transport::session::SharedState;
 use crate::{Consumer, Producer};
 
 pub struct Session {
@@ -10,9 +11,9 @@ pub struct Session {
 }
 
 impl Session {
-    pub async fn run(self) -> Result<(), SessionError> {
+    pub async fn run(self, shared_state: SharedState) -> Result<(), SessionError> {
         let mut tasks = FuturesUnordered::new();
-        tasks.push(self.session.run().boxed());
+        tasks.push(self.session.run(shared_state.clone()).boxed());
 
         if let Some(producer) = self.producer {
             tasks.push(producer.run().boxed());
@@ -21,7 +22,6 @@ impl Session {
         if let Some(consumer) = self.consumer {
             tasks.push(consumer.run().boxed());
         }
-
         tasks.select_next_some().await
     }
 }