+++ moq-pub/CHANGELOG.md
@@ -6,6 +6,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.8.5](https://github.com/rimmelb/moq-rs/compare/moq-pub-v0.8.4...moq-pub-v0.8.5) - 2025-04-27
+
+### Other
+
+- Fix
+- Fix as.ref instead of to_string
+- Fix some errors concerning the removal of moq-shared
+- Deleted moq-shared and moved to moq_transport::session
+- GoAway function-nel kiegészített kód
+
 ## [0.8.4](https://github.com/englishm/moq-rs/compare/moq-pub-v0.8.3...moq-pub-v0.8.4) - 2025-02-24
 
 ### Other