+++ moq-sub/CHANGELOG.md
@@ -6,6 +6,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.4](https://github.com/rimmelb/moq-rs/compare/moq-sub-v0.3.3...moq-sub-v0.3.4) - 2025-04-27
+
+### Other
+
+- Fix
+- Fix as.ref instead of to_string
+- Fix some errors concerning the removal of moq-shared
+- Deleted moq-shared and moved to moq_transport::session
+- Időzítő megadható az URL-ben is
+- GoAway function-nel kiegészített kód
+
 ## [0.3.3](https://github.com/englishm/moq-rs/compare/moq-sub-v0.3.2...moq-sub-v0.3.3) - 2025-02-24
 
 ### Other